+++ rostests/winetests/kernel32/profile.c
@@ -92,12 +92,12 @@ static void test_profile_int(void)
 
 static void test_profile_string(void)
 {
-    static WCHAR emptyW[] = { 0 };
-    static WCHAR keyW[] = { 'k','e','y',0 };
-    static WCHAR sW[] = { 's',0 };
-    static WCHAR TESTFILE2W[] = {'.','\\','t','e','s','t','w','i','n','e','2','.','i','n','i',0};
-    static WCHAR valsectionW[] = {'v','a','l','_','e','_','s','e','c','t','i','o','n',0 };
-    static WCHAR valnokeyW[] = {'v','a','l','_','n','o','_','k','e','y',0};
+    static WCHAR emptyW[] = { 0 }; /* if "const", GetPrivateProfileStringW(emptyW, ...) crashes on win2k */
+    static const WCHAR keyW[] = { 'k','e','y',0 };
+    static const WCHAR sW[] = { 's',0 };
+    static const WCHAR TESTFILE2W[] = {'.','\\','t','e','s','t','w','i','n','e','2','.','i','n','i',0};
+    static const WCHAR valsectionW[] = {'v','a','l','_','e','_','s','e','c','t','i','o','n',0 };
+    static const WCHAR valnokeyW[] = {'v','a','l','_','n','o','_','k','e','y',0};
     HANDLE h;
     int ret;
     DWORD count;
@@ -146,13 +146,15 @@ static void test_profile_string(void)
     ret=GetPrivateProfileStringW(emptyW, keyW, emptyW, bufW,
                                  sizeof(bufW)/sizeof(bufW[0]), TESTFILE2W);
     todo_wine
+    ok(ret == 13, "expected 13, got %u\n", ret);
+    todo_wine
     ok(!lstrcmpW(valsectionW,bufW), "expected %s, got %s\n",
         wine_dbgstr_w(valsectionW), wine_dbgstr_w(bufW) );
 
     /* works only in unicode, ascii crashes */
     ret=GetPrivateProfileStringW(sW, emptyW, emptyW, bufW,
                                  sizeof(bufW)/sizeof(bufW[0]), TESTFILE2W);
-    todo_wine
+    ok(ret == 10, "expected 10, got %u\n", ret);
     ok(!lstrcmpW(valnokeyW,bufW), "expected %s, got %s\n",
         wine_dbgstr_w(valnokeyW), wine_dbgstr_w(bufW) );
 
@@ -378,19 +380,19 @@ static void test_profile_existing(void)
 
     for (i=0; i < sizeof(pe)/sizeof(pe[0]); i++)
     {
-        h = CreateFile(testfile1, pe[i].dwDesiredAccess, pe[i].dwShareMode, NULL,
+        h = CreateFileA(testfile1, pe[i].dwDesiredAccess, pe[i].dwShareMode, NULL,
                        CREATE_ALWAYS, FILE_ATTRIBUTE_NORMAL, NULL);
         ok(INVALID_HANDLE_VALUE != h, "%d: CreateFile failed\n",i);
         SetLastError(0xdeadbeef);
 
-        ret = WritePrivateProfileString(SECTION, KEY, "12345", testfile1);
+        ret = WritePrivateProfileStringA(SECTION, KEY, "12345", testfile1);
         if (!pe[i].write_error)
         {
             if (!ret)
                 ok( broken(GetLastError() == pe[i].broken_error),
                     "%d: WritePrivateProfileString failed with error %u\n", i, GetLastError() );
             CloseHandle(h);
-            size = GetPrivateProfileString(SECTION, KEY, 0, buffer, MAX_PATH, testfile1);
+            size = GetPrivateProfileStringA(SECTION, KEY, 0, buffer, MAX_PATH, testfile1);
             if (ret)
                 ok( size == 5, "%d: test failed, number of characters copied: %d instead of 5\n", i, size );
             else
@@ -404,21 +406,21 @@ static void test_profile_existing(void)
                 ok( err == pe[i].write_error, "%d: WritePrivateProfileString failed with error %u/%u\n",
                     i, err, pe[i].write_error );
             CloseHandle(h);
-            size = GetPrivateProfileString(SECTION, KEY, 0, buffer, MAX_PATH, testfile1);
+            size = GetPrivateProfileStringA(SECTION, KEY, 0, buffer, MAX_PATH, testfile1);
             ok( !size, "%d: test failed, number of characters copied: %d instead of 0\n", i, size );
         }
 
-        ok( DeleteFile(testfile1), "delete failed\n" );
+        ok( DeleteFileA(testfile1), "delete failed\n" );
     }
 
-    h = CreateFile(testfile2, GENERIC_WRITE, 0, NULL, CREATE_ALWAYS, FILE_ATTRIBUTE_NORMAL, NULL);
+    h = CreateFileA(testfile2, GENERIC_WRITE, 0, NULL, CREATE_ALWAYS, FILE_ATTRIBUTE_NORMAL, NULL);
     sprintf( buffer, "[%s]\r\n%s=123\r\n", SECTION, KEY );
     ok( WriteFile( h, buffer, strlen(buffer), &size, NULL ), "failed to write\n" );
     CloseHandle( h );
 
     for (i=0; i < sizeof(pe)/sizeof(pe[0]); i++)
     {
-        h = CreateFile(testfile2, pe[i].dwDesiredAccess, pe[i].dwShareMode, NULL,
+        h = CreateFileA(testfile2, pe[i].dwDesiredAccess, pe[i].dwShareMode, NULL,
                        OPEN_EXISTING, FILE_ATTRIBUTE_NORMAL, NULL);
         ok(INVALID_HANDLE_VALUE != h, "%d: CreateFile failed\n",i);
         SetLastError(0xdeadbeef);
@@ -432,24 +434,24 @@ static void test_profile_existing(void)
             ok( !ret, "%d: GetPrivateProfileString succeeded\n", i );
         CloseHandle(h);
     }
-    ok( DeleteFile(testfile2), "delete failed\n" );
+    ok( DeleteFileA(testfile2), "delete failed\n" );
 }
 
 static void test_profile_delete_on_close(void)
 {
-    static CHAR testfile[] = ".\\testwine5.ini";
     HANDLE h;
     DWORD size, res;
+    static const CHAR testfile[] = ".\\testwine5.ini";
     static const char contents[] = "[" SECTION "]\n" KEY "=123\n";
 
-    h = CreateFile(testfile, GENERIC_WRITE, FILE_SHARE_READ, NULL,
+    h = CreateFileA(testfile, GENERIC_WRITE, FILE_SHARE_READ, NULL,
                     CREATE_ALWAYS, FILE_FLAG_DELETE_ON_CLOSE, NULL);
-    ok( WriteFile( h, contents, sizeof contents - 1, &size, NULL ),
-                    "Cannot write test file: %x\n", GetLastError() );
+    res = WriteFile( h, contents, sizeof contents - 1, &size, NULL );
+    ok( res, "Cannot write test file: %x\n", GetLastError() );
     ok( size == sizeof contents - 1, "Test file: partial write\n");
 
     SetLastError(0xdeadbeef);
-    res = GetPrivateProfileInt(SECTION, KEY, 0, testfile);
+    res = GetPrivateProfileIntA(SECTION, KEY, 0, testfile);
     ok( res == 123 ||
         broken(res == 0 && GetLastError() == ERROR_SHARING_VIOLATION), /* Win9x, WinME */
         "Got %d instead of 123\n", res);
@@ -460,20 +462,20 @@ static void test_profile_delete_on_close(void)
 
 static void test_profile_refresh(void)
 {
-    static CHAR testfile[] = ".\\winetest4.ini";
+    static const CHAR testfile[] = ".\\winetest4.ini";
     HANDLE h;
     DWORD size, res;
     static const char contents1[] = "[" SECTION "]\n" KEY "=123\n";
     static const char contents2[] = "[" SECTION "]\n" KEY "=124\n";
 
-    h = CreateFile(testfile, GENERIC_WRITE, FILE_SHARE_READ, NULL,
+    h = CreateFileA(testfile, GENERIC_WRITE, FILE_SHARE_READ, NULL,
                     CREATE_ALWAYS, FILE_FLAG_DELETE_ON_CLOSE, NULL);
-    ok( WriteFile( h, contents1, sizeof contents1 - 1, &size, NULL ),
-                    "Cannot write test file: %x\n", GetLastError() );
+    res = WriteFile( h, contents1, sizeof contents1 - 1, &size, NULL );
+    ok( res, "Cannot write test file: %x\n", GetLastError() );
     ok( size == sizeof contents1 - 1, "Test file: partial write\n");
 
     SetLastError(0xdeadbeef);
-    res = GetPrivateProfileInt(SECTION, KEY, 0, testfile);
+    res = GetPrivateProfileIntA(SECTION, KEY, 0, testfile);
     ok( res == 123 ||
         broken(res == 0 && GetLastError() == ERROR_SHARING_VIOLATION), /* Win9x, WinME */
         "Got %d instead of 123\n", res);
@@ -482,20 +484,27 @@ static void test_profile_refresh(void)
 
     /* Test proper invalidation of wine's profile file cache */
 
-    h = CreateFile(testfile, GENERIC_WRITE, FILE_SHARE_READ, NULL,
+    h = CreateFileA(testfile, GENERIC_WRITE, FILE_SHARE_READ, NULL,
                     CREATE_ALWAYS, FILE_FLAG_DELETE_ON_CLOSE, NULL);
-    ok( WriteFile( h, contents2, sizeof contents2 - 1, &size, NULL ),
-                    "Cannot write test file: %x\n", GetLastError() );
+    res = WriteFile( h, contents2, sizeof contents2 - 1, &size, NULL );
+    ok( res, "Cannot write test file: %x\n", GetLastError() );
     ok( size == sizeof contents2 - 1, "Test file: partial write\n");
 
     SetLastError(0xdeadbeef);
-    res = GetPrivateProfileInt(SECTION, KEY, 0, testfile);
+    res = GetPrivateProfileIntA(SECTION, KEY, 0, testfile);
     ok( res == 124 ||
         broken(res == 0 && GetLastError() == 0xdeadbeef), /* Win9x, WinME */
         "Got %d instead of 124\n", res);
 
     /* This also deletes the file */
     CloseHandle(h);
+
+    /* Cache must be invalidated if file no longer exists and default must be returned */
+    SetLastError(0xdeadbeef);
+    res = GetPrivateProfileIntA(SECTION, KEY, 421, testfile);
+    ok( res == 421 ||
+        broken(res == 0 && GetLastError() == 0xdeadbeef), /* Win9x, WinME */
+        "Got %d instead of 421\n", res);
 }
 
 static void create_test_file(LPCSTR name, LPCSTR data, DWORD size)
@@ -1008,18 +1017,22 @@ static void test_WritePrivateProfileString(void)
        broken(GetLastError() == ERROR_PATH_NOT_FOUND), /* Win9x and WinME */
        "Expected ERROR_ACCESS_DENIED, got %d\n", GetLastError());
 
-    /* The resulting file will be  X:\\%WINDIR%\\win1.tmp */
+    /* Relative paths are relative to X:\\%WINDIR% */
     GetWindowsDirectoryA(temp, MAX_PATH);
     GetTempFileNameA(temp, "win", 1, path);
-    DeleteFileA(path);
+    if (GetFileAttributesA(path) == INVALID_FILE_ATTRIBUTES)
+        skip("Not allowed to create a file in the Windows directory\n");
+    else
+    {
+        DeleteFileA(path);
 
-    /* relative path in lpFileName */
-    data = "[App]\r\n"
-           "key=string\r\n";
-    ret = WritePrivateProfileStringA("App", "key", "string", "win1.tmp");
-    ok(ret == TRUE, "Expected TRUE, got %d\n", ret);
-    ok(check_file_data(path, data), "File doesn't match\n");
-    DeleteFileA(path);
+        data = "[App]\r\n"
+               "key=string\r\n";
+        ret = WritePrivateProfileStringA("App", "key", "string", "win1.tmp");
+        ok(ret == TRUE, "Expected TRUE, got %d, le=%u\n", ret, GetLastError());
+        ok(check_file_data(path, data), "File doesn't match\n");
+        DeleteFileA(path);
+    }
 
     GetTempPathA(MAX_PATH, temp);
     GetTempFileNameA(temp, "wine", 0, path);
